requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
dj-stripe==2.9.1
Django==5.1.2
django-cors-headers==4.9.0
django-environ==0.11.2
django-grappelli==4.0.1
djangorestframework==3.16.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
gunicorn==22.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.1
mysql-connector-python==9.1.0
mysqlclient==2.2.7
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlearn==0.0.1
sqlparse==0.5.1
stripe==11.6.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
whitenoise==6.7.0
